Creates an error modal when session is not found
Creates an error modal for when a user goes to the emulator with a
previous session but the session cannot be found on the server. Previously
what would happen is the frontend would continually try to get the
session without telling the user there is a problem. This also provides
a button which the user can press which resets the session in order
to create a new toynet session.